% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_vs_del.R
\name{metric_vs_del}
\alias{metric_vs_del}
\title{Robustness Simulation: Metric vs. Vertex Deletions}
\usage{
metric_vs_del(
  metric_fun,
  ...,
  sim = NULL,
  graph = NULL,
  del_attrs = c("random", "degree", "pagerank", "harmonic", "close", "between", "eigen"),
  del_min = 0,
  del_max = 0.5,
  nchunks = 10,
  plot = TRUE,
  seed_rand = NULL,
  recalc_VDOs = TRUE,
  pass_graph = TRUE,
  reuse_dists = FALSE
)
}
\arguments{
\item{reuse_dists}{}
}
\description{
Robustness Simulation: Metric vs. Vertex Deletions
}
