% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Ys_vs_X.R
\name{plot_Ys_vs_X}
\alias{plot_Ys_vs_X}
\title{Plot multiple Y columns vs. one X column}
\usage{
plot_Ys_vs_X(
  df,
  x_col_name = "del_frac",
  x_axis_label = "Vertices Deleted (Fraction of Total)",
  y_axis_label = "Y",
  title = sprintf("\%s vs. \%s", y_axis_label, x_axis_label),
  legend_title = "VDO Attribute",
  return_plot = FALSE
)
}
\arguments{
\item{return_plot}{}
}
\description{
Plot multiple Y columns vs. one X column
}
