% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_class.R
\name{require_class}
\alias{require_class}
\title{require_class}
\usage{
require_class(req_class, obj, obj_name)
}
\arguments{
\item{req_class}{\code{(character)} : the required class name.}

\item{obj}{\verb{(R object)} :  the object.}

\item{obj_name}{\code{(character)} : the name of the object in memory.}
}
\description{
Require an object to have the specified class.  If it doesn't, an error is thrown.
}
